requirements.txt
google-generativeai
youtube-transcript-api
streamlit
python-dotenv